Switched repository primary key from Integer to Long

+++ src/main/java/com/mellora/drugconsortium/dao/EmployeeRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.mellora.drugconsortium.model.Employee;
 
-public interface EmployeeRepo extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepo extends JpaRepository<Employee, Long> {
 	@Query(value = "SELECT * FROM employee e WHERE e.FIRST_NAME = :first AND e.LAST_NAME = :last", nativeQuery = true)
 	public Employee findEmployeeByFirstAndLastName(@Param("first") String firstName, @Param("last") String lastName);
 }
